feature: migrated back to c file, and added procps list processes
so after @HUBLACKBE said to check again if there want any library we
could use. I found procps and i bassicly does what i already was doing,
reading the files and the status file and checks the state and a couple
of other things. The problem is that it needs an external package and
thus creating an external dependency. we want the program to be
lightweight. this would save me time but would also cost me the
simplicity of the program.